#11 closed: Change commandId types from uint to byte

+++ src/SimpleSerialProtocol.h
@@ -13,21 +13,21 @@ const unsigned long COMMAND_MESSAGE_COMPLETE_TIMEOUT = 1000;
 
 const byte CHAR_EOT = 0x0A; // 10 // End of Transmission - Line Feed Zeichen \n
 
-const unsigned int COMMAND_CALLBACK_RANGE_FROM = 0;
-const unsigned int COMMAND_CALLBACK_RANGE_TO = 255;
+const byte COMMAND_CALLBACK_RANGE_FROM = 0;
+const byte COMMAND_CALLBACK_RANGE_TO = 255;
 
-const unsigned int ERROR_IS_DEAD = 255;
-const unsigned int ERROR_EOT_WAS_NOT_READ = 254;
-const unsigned int ERROR_IS_NOT_WAITING_FOR_READ_EOT = 253;
-const unsigned int ERROR_IS_NOT_EOT = 252;
+const uint8_t ERROR_IS_DEAD = 255;
+const uint8_t ERROR_EOT_WAS_NOT_READ = 254;
+const uint8_t ERROR_IS_NOT_WAITING_FOR_READ_EOT = 253;
+const uint8_t ERROR_IS_NOT_EOT = 252;
 
-const unsigned int ERROR_WAIT_FOR_BYTE_TIMEOUT = 0;
-const unsigned int ERROR_IS_NOT_INITIALIZED = 1;
-const unsigned int ERROR_COMMAND_RANGE_IS_INVALID = 2;
-const unsigned int ERROR_COMMAND_IS_NOT_IN_RESERVED_RANGE = 3;
-const unsigned int ERROR_COMMAND_IS_REGISTERED = 4;
-const unsigned int ERROR_COMMAND_IS_NOT_REGISTERED = 5;
-const unsigned int ERROR_END_OF_STRING_BYTE_NOT_IN_CHAR_ARRAY = 6;
+const uint8_t ERROR_WAIT_FOR_BYTE_TIMEOUT = 0;
+const uint8_t ERROR_IS_NOT_INITIALIZED = 1;
+const uint8_t ERROR_COMMAND_RANGE_IS_INVALID = 2;
+const uint8_t ERROR_COMMAND_IS_NOT_IN_RESERVED_RANGE = 3;
+const uint8_t ERROR_COMMAND_IS_REGISTERED = 4;
+const uint8_t ERROR_COMMAND_IS_NOT_REGISTERED = 5;
+const uint8_t ERROR_END_OF_STRING_BYTE_NOT_IN_CHAR_ARRAY = 6;
 
 
 class SimpleSerialProtocol : public Core {
@@ -35,7 +35,7 @@ class SimpleSerialProtocol : public Core {
 public:
     typedef void (*CallbackPointer)();
 
-    typedef void (*ErrorCallbackPointer)(unsigned int errorNum);
+    typedef void (*ErrorCallbackPointer)(uint8_t errorNum);
 
 
 #ifdef SOFTWARESERIAL_SUPPORTED
@@ -45,16 +45,16 @@ public:
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO);
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO);
 
     SimpleSerialProtocol(
             SoftwareSerial *softwareSerialPtr,
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO);
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO);
 
 #endif
 
@@ -63,8 +63,8 @@ public:
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
     );
 
     SimpleSerialProtocol(
@@ -72,28 +72,28 @@ public:
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
     );
 
     SimpleSerialProtocol(
             Stream &streamRef,
-            unsigned int streamType,
+            uint8_t streamType,
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
     );
 
     SimpleSerialProtocol(
             Stream *streamPtr,
-            unsigned int streamType,
+            uint8_t streamType,
             unsigned long baudrate,
             unsigned long waitForByteTimeout,
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
-            unsigned int commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
+            byte commandCallbackRangeFrom = COMMAND_CALLBACK_RANGE_FROM,
+            byte commandCallbackRangeTo = COMMAND_CALLBACK_RANGE_TO
     );
 
     ~SimpleSerialProtocol();
@@ -118,8 +118,8 @@ public:
 
 protected:
 
-    unsigned int commandCallbackRangeFrom;
-    unsigned int commandCallbackRangeTo;
+    byte commandCallbackRangeFrom;
+    byte commandCallbackRangeTo;
 
     bool _isInitialized = false;
     bool _isDead = false;
@@ -134,10 +134,10 @@ protected:
 
     void registerCommandCallback(byte command, CallbackPointer commandCallbackPointer); //
 
-    unsigned int getCommandIndex(byte command); //
+    uint8_t getCommandIndex(byte command); //
     void callCommandCallback(byte command);
 
-    void error(unsigned int errorNum, bool dieInstantly); //
+    void error(uint8_t errorNum, bool dieInstantly); //
     void flushCommand();
 
 private:
@@ -149,8 +149,8 @@ private:
 
     void afterConstructor(
             ErrorCallbackPointer errorCallbackPointer,
-            unsigned int commandCallbackRangeFrom,
-            unsigned int commandCallbackRangeTo
+            byte commandCallbackRangeFrom,
+            byte commandCallbackRangeTo
     );
 
     void die();
